Move transport-related tests to a new file
